+++ service/src/main/java/com/noqapp/service/QueueService.java
@@ -648,7 +648,7 @@ public class QueueService {
         queueManager.updateCustomerPriorityAndCustomerAttributes(qid, codeQR, tokenNumber, customerPriorityLevel, businessCustomerAttribute);
     }
 
-    public long countDistinctQIDsInBiz(String bizNameId, int days) {
+    public int countDistinctQIDsInBiz(String bizNameId, int days) {
         return queueManagerJDBC.countDistinctQIDsInBiz(bizNameId, days);
     }
